initial split bofhbot to bofhbot_lib.  untested
